requirements.txt
Pillow==9.1.1
PySide6==6.2.4
shiboken6==6.2.4
